+++ crates/engine/src/debug/stacktrace.rs
@@ -1,3 +1,8 @@
+use alloc::{
+    boxed::Box,
+    string::{String, ToString},
+    vec::Vec,
+};
 use std::{
     borrow::Cow,
     cell::OnceCell,