+++ crates/engine/src/exec/dap.rs
@@ -1,6 +1,12 @@
+use alloc::{
+    borrow::ToOwned,
+    string::{String, ToString},
+    vec::Vec,
+};
 use std::{
     cell::RefCell,
     collections::{BTreeMap, BTreeSet},
+    eprintln,
     io::{BufReader, BufWriter},
     net::TcpListener,
     path::{Path, PathBuf},